combined  duck script and move script
