--- Assets/Scenes/EnemyMovementTestScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ceb28fbaa5b6e4340aa452e2efd1305f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
