--- Assets/Scripts/Mechanics/WesTestScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 93b0738d5db6f9047be79c770c4d349f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
