--- Assets/Scripts/Mechanics/WesTestScripts/DuckScript.cs
@@ -1,27 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class DuckScript : MonoBehaviour
-{
-    
-    void Start()
-    {
-        
-    }
-
-    
-    void Update()
-    {
-        if (Input.GetMouseButtonDown(0))
-        {
-            Vector2 mousePosition = Camera.main.ScreenToWorldPoint(Input.mousePosition);
-            if (GetComponent<Collider2D>().OverlapPoint(mousePosition))
-            {
-                GetComponent<Rigidbody2D>().velocity = new Vector2(0, -5);
-            }
-        }
-    }
-
-}
