--- Assets/Scripts/Mechanics/WesTestScripts/DuckScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3a7d0a87d8ef5fb40ab23e480650842a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
